+++ src/main/java/mod/gottsch/forge/treasure2/core/setup/CommonSetup.java
@@ -321,6 +321,7 @@ public class CommonSetup {
 		TreasureApi.registerChest(TreasureBlocks.CARDBOARD_BOX);
 		TreasureApi.registerChest(TreasureBlocks.MILK_CRATE);
 		TreasureApi.registerChest(TreasureBlocks.BARREL_CHEST);
+		TreasureApi.registerChest(TreasureBlocks.VANILLA_CHEST);
 		TreasureApi.registerChest(TreasureBlocks.WITHER_CHEST);
 
 		// register mimics
@@ -333,6 +334,7 @@ public class CommonSetup {
 		TreasureApi.registerMimic(TreasureBlocks.CARDBOARD_BOX.getId(), TreasureEntities.CARDBOARD_BOX_MIMIC_ENTITY_TYPE.getId());
 		TreasureApi.registerMimic(TreasureBlocks.MILK_CRATE.getId(), TreasureEntities.MILK_CRATE_MIMIC_ENTITY_TYPE.getId());
 		TreasureApi.registerMimic(TreasureBlocks.BARREL_CHEST.getId(), TreasureEntities.BARREL_MIMIC_ENTITY_TYPE.getId());
+		TreasureApi.registerMimic(TreasureBlocks.VANILLA_CHEST.getId(), TreasureEntities.VANILLA_CHEST_MIMIC_ENTITY_TYPE.getId());
 
 		/*
 		 *  register wishable handlers
@@ -453,7 +455,8 @@ public class CommonSetup {
 		event.put(TreasureEntities.MOLDY_CRATE_CHEST_MIMIC_ENTITY_TYPE.get(), MoldyCrateChestMimic.createAttributes().build());
 		event.put(TreasureEntities.CARDBOARD_BOX_MIMIC_ENTITY_TYPE.get(), CardboardBoxMimic.createAttributes().build());
 		event.put(TreasureEntities.MILK_CRATE_MIMIC_ENTITY_TYPE.get(), MilkCrateMimic.createAttributes().build());
-		event.put(TreasureEntities.BARREL_MIMIC_ENTITY_TYPE.get(), MilkCrateMimic.createAttributes().build());
+		event.put(TreasureEntities.BARREL_MIMIC_ENTITY_TYPE.get(), BarrelMimic.createAttributes().build());
+		event.put(TreasureEntities.VANILLA_CHEST_MIMIC_ENTITY_TYPE.get(), VanillaChestMimic.createAttributes().build());
 	}
 
 	@SubscribeEvent
@@ -471,6 +474,6 @@ public class CommonSetup {
 		event.register(TreasureEntities.CARDBOARD_BOX_MIMIC_ENTITY_TYPE.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, Mob::checkMobSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
 		event.register(TreasureEntities.MILK_CRATE_MIMIC_ENTITY_TYPE.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, Mob::checkMobSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
 		event.register(TreasureEntities.BARREL_MIMIC_ENTITY_TYPE.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, Mob::checkMobSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
-
+		event.register(TreasureEntities.VANILLA_CHEST_MIMIC_ENTITY_TYPE.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, Mob::checkMobSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
 	}
 }